Add files via upload
Beginning with basic data cleaning on Excel and moving on to the analysis phase on Microsoft SQL Server, I noticed the lack of data in certain fields, specifically the ones that required user input to log any undetectable changes from the fitness tracker device. Visualized my findings on Tableau and placed my recommendations in the word document.